Diplom_avtotest_Chitai_gorod/requirements.txt
selenium>=4.0.0
requests>=2.26.0
pytest>=7.0.0
allure-pytest>=2.9.0
pytest-xdist>=2.5.0
pylint>=2.12.0
flake8>=4.0.0